feature/req-180326

+++ src/main/java/com/svc/ventas/message/response/ProductoSearchResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -23,7 +25,7 @@ public class ProductoSearchResponse {
 
 	private String imagen;
 	
-	private float precioBase;
+	private BigDecimal precioBase;
 
 	private Boolean estado;
 }
